Organize things a bit
